requirements.txt
torch>=1.9.0
torchvision>=0.10.0
open_clip_torch>=2.0.0
Pillow>=8.0.0
numpy>=1.20.0
matplotlib>=3.3.0
opencv-python>=4.5.0
tqdm>=4.60.0
ultralytics>=8.0.0
scipy>=1.7.0
fastapi>=0.68.0
uvicorn[standard]>=0.15.0
websockets>=10.0
python-multipart>=0.0.5 